Added VPD
This is a major change - I added two new variables to the analysis, min and max Vapor Pressure Deficit, and the associated input and output files. I also changed how years were considered for the analysis (now it is 20 years before the first visit and 10 yrs between visits).  The Shiny code has been updated. The changes are not complete.